+++ components/divider.tsx
@@ -1,5 +1,6 @@
 /** @jsxImportSource theme-ui */
 import { ThemeUIStyleObject } from 'theme-ui'
+import { Row, Column } from '@carbonplan/components'
 
 interface DividerProps {
   color?: string
@@ -16,5 +17,17 @@ export function Divider({ color = 'gray', sx = {} }: DividerProps) {
         ...sx,
       }}
     />
+  )
+}
+
+export const SectionDivider: React.FC<{ color: string }> = ({ color }) => {
+  return (
+    <>
+      <Row columns={[12]}>
+        <Column start={[2]} width={[10]}>
+          <Divider color={color} />
+        </Column>
+      </Row>
+    </>
   )
 }
